refactor(admin): modularize Forge run orchestration
